pyproject.toml
[tool.poetry]
name = "colorama-experiment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
langchain = "0.1.9"
huggingface-hub = "^0.24.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
